Refactor the DataManager in its own file
